requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
pandas==2.1.1
loguru==0.7.2
python-dotenv==1.0.0